requirements.txt
Django==4.1.7
django-admin-interface==0.24.2
django-admin-rangefilter==0.10.0
django-autocompletefilter==0.0.8
django-axes==5.40.1
django-celery-beat==2.6.0
django-celery-results==2.4.0
django-colorfield==0.8.0
django-crontab==0.7.1
django-debug-toolbar==3.8.1
django-filter==23.1
django-flat-theme==1.1.4
django-import-export==2.8.0
django-json-widget==1.1.1
django-multi-captcha-admin==2.0.0
django-recaptcha==3.0.0
django-seed==0.2.2
django-simple-history==3.0.0
django-storages==1.13.1
djangorestframework==3.13.1
gunicorn==20.1.0
mysqlclient==2.1.1
onesignal-sdk==2.0.0
django-tinymce
celery==5.2.7
django-celery-beat==2.6.0
redis==3.5.3
django-celery-results==2.4.0
kombu==5.2.4
gevent==22.10.1
pika